Detangle CC serialization API from `:core` and move it `:object-graph-serialization`
