+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/configurationcache/metadata/ProjectMetadataController.kt
@@ -28,7 +28,7 @@ import org.gradle.configurationcache.ConfigurationCacheStateStore
 import org.gradle.configurationcache.DefaultConfigurationCache
 import org.gradle.configurationcache.StateType
 import org.gradle.configurationcache.models.ProjectStateStore
-import org.gradle.configurationcache.serialization.IsolateOwner
+import org.gradle.configurationcache.serialization.IsolateOwners
 import org.gradle.configurationcache.serialization.ReadContext
 import org.gradle.configurationcache.serialization.WriteContext
 import org.gradle.configurationcache.serialization.ownerService
@@ -73,7 +73,7 @@ class ProjectMetadataController(
 
     override fun write(encoder: Encoder, value: LocalComponentGraphResolveState) {
         val (context, codecs) = cacheIO.writerContextFor(encoder)
-        context.push(IsolateOwner.OwnerHost(host), codecs.userTypesCodec())
+        context.push(IsolateOwners.OwnerHost(host), codecs.userTypesCodec())
         context.runWriteOperation {
             write(value.id)
             write(value.moduleVersionId)
@@ -128,7 +128,7 @@ class ProjectMetadataController(
 
     override fun read(decoder: Decoder): LocalComponentGraphResolveState {
         val (context, codecs) = cacheIO.readerContextFor(decoder)
-        context.push(IsolateOwner.OwnerHost(host), codecs.userTypesCodec())
+        context.push(IsolateOwners.OwnerHost(host), codecs.userTypesCodec())
         return context.runReadOperation {
             val id = readNonNull<ComponentIdentifier>()
             val moduleVersionId = readNonNull<ModuleVersionIdentifier>()
