--- platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/configurationcache/serialization/beans/BeanStateReader.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright 2020 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.configurationcache.serialization.beans
-
-import org.gradle.configurationcache.serialization.ReadContext
-
-
-interface BeanStateReader {
-
-    fun ReadContext.newBeanWithId(generated: Boolean, id: Int) =
-        newBean(generated).also {
-            isolate.identities.putInstance(id, it)
-        }
-
-    fun ReadContext.newBean(generated: Boolean): Any
-
-    suspend fun ReadContext.readStateOf(bean: Any)
-}
